Part 2: Removed OnyxTelemetry file and moved code into OnyxClient
